Update enemy class and textures

+++ Czolgi/player.h
@@ -18,8 +18,8 @@ public:
 
 private:
     QString name;
-    int angle = 0;
-    int speed = 5;
+    qreal angle = 0;
+    qreal speed = 5;
 };
 
 #endif // PLAYER_H
